LandIce: Introducing a StokesFOBase class
All other StokesFO* problems should aim to inherit from this, to limit
the amount of code duplication and the consequent risk of divergence
between the velocity part of different models.
