+++ src/LandIce/problems/LandIce_StokesFOThermoCoupled.cpp
@@ -388,7 +388,6 @@ LandIce::StokesFOThermoCoupled::getValidProblemParameters () const
   validPL->sublist("LandIce Field Norm", false, "");
   validPL->sublist("LandIce Physical Parameters", false, "");
   validPL->sublist("LandIce Noise", false, "");
-  validPL->sublist("Parameter Fields", false, "Parameter Fields to be registered");
   validPL->set<bool>("Use Time Parameter", false, "Solely to use Solver Method = Continuation");
   validPL->set<bool>("Print Stress Tensor", false, "Whether to save stress tensor in the mesh");
